% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_method.R
\name{BinarizeSignaturePrediction}
\alias{BinarizeSignaturePrediction}
\title{Get BinarySignature Prediction}
\usage{
BinarizeSignaturePrediction(
  x.train,
  method = c("greedy", "threshold"),
  ground.truth = NULL,
  x.threshold = NULL,
  x.test = NULL,
  weights = NULL
)
}
\arguments{
\item{x.train}{numerical vector; The training score vector}

\item{method}{character; The threhsold computation method. Possible values 
are "greedy" and "threshold".
Default = "greedy"}

\item{ground.truth}{logical vector; The logical ground truth needed for when method = "greedy"}

\item{x.threshold}{numerical; The threshold input for when method = "threshold"}

\item{x.test}{numerical vector; The testing score vector. If NULL, return no testing prediction
Default = NULL}

\item{weights}{numerical vector of length identical to x.train; The observation weights.
Default = NULL i.e. 1 for each observation}
}
\value{
list("y.train.pred" = ..., "y.test.pred" = ...)
}
\description{
Function to get the prediction for some numerical score. In brief, a threshold 
is applied on a score vector to get binary prediction which are returned by 
the function
The threshold can be provided by the user (method = "threshold") or the 
threshold can be computed in a "greedy" manner by optimizing the accuracy. 
So there are two methods: 
- "threshold": unbiased method where `x.threshold` must be provided
- "greedy": Optimal threshold is computed to optimize maxium accuracy. Here, 
'ground.truth' needs to be provided.
}
