% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross_Validation.R
\name{CreateCVFolds}
\alias{CreateCVFolds}
\title{Create Cross-validation folds}
\usage{
CreateCVFolds(
  y,
  y.leave.1.out = NULL,
  k = 5,
  replicates = 1,
  sampling.method = c("under", "upper", "none"),
  seed = .Random.seed[1]
)
}
\arguments{
\item{y}{vector; the vector to create the CV folds from}

\item{y.leave.1.out}{vector; If not NULL, vector to create Leave-1-out category CV}

\item{k}{numerical: The number of outer folds (default = 5)}

\item{replicates}{numerical; The number of replicates. At each replicate, the seed is increased by 1 (default = 1)}

\item{sampling.method}{character; The sampling method to balance the folds:
- "under": Undersampling is performed to balance the categories in every fold (default)
- "upper": Oversampling is performed to balance the categories in every fold
- "none": No balancing sampling is performed}

\item{seed}{numerical: The seed (default = .Random.seed[1])}
}
\value{
list containing the index of each sample belonging to the training and testing set for every fold
  Names of the list correspond to the names of CV folds, and values of the list contain two names:
  - "train": Training sample indexes
  - "test": Testing sample indexes
}
\description{
Function to create cross-validation folds
}
