% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross_Validation.R
\name{CreateNestedCVFolds}
\alias{CreateNestedCVFolds}
\title{Create Nested-Cross-validation folds}
\usage{
CreateNestedCVFolds(
  y,
  y.leave.1.out = NULL,
  k.out = 5,
  k.in = 5,
  replicates = 1,
  sampling.method = c("under", "upper", "none"),
  seed = .Random.seed[1]
)
}
\arguments{
\item{y}{vector; the vector to create the CV folds from}

\item{y.leave.1.out}{vector; If not NULL, vector to create Leave-1-out category CV}

\item{k.out}{numerical: The number of outer folds (default = 5)}

\item{k.in}{numerical: The number of inner folds (default = 5)}

\item{replicates}{numerical; The number of replicates. At each replicate, the seed is increased by 1}

\item{sampling.method}{character; The sampling method to balance the folds:
- "under": Undersampling is performed to balance the categories in every fold (default)
- "upper": Oversampling is performed to balance the categories in every fold
- "none": No balancing sampling is performed}

\item{seed}{numerical: The seed (default = .Random.seed[1])}
}
\value{
list of lists. The first level corresponds to outer folds, and the second level contains inner folds.
The content of the inner-fold list is the same as in CreateCVFolds().
}
\description{
Function to create nested cross-validation folds using the CreateCVFolds in this package.
See ?CreateCVFolds() for more details.
}
