% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross_Validation.R
\name{CrossValidation}
\alias{CrossValidation}
\title{Cross Validation}
\usage{
CrossValidation(
  x,
  y,
  CV.k = 5,
  CV.y.leave.1.out.column = NULL,
  CV.replicates = 1,
  CV.sampling.method = c("under", "upper", "none"),
  y.label,
  y.covariates = NULL,
  y.sample = NULL,
  method = c("LR", "signature"),
  hyperparams = NULL,
  data.trans.method = c("none", FEATURE.TRANS.METHODS),
  pca.explained.var.threshold = NULL,
  rm.corr.features = F,
  DA.method = c("none", DEA.METHODS),
  DA.p.adjust.method = P.ADJUST.METHODS,
  DA.p.val.threshold = 0.05,
  DA.event.per.variable = NULL,
  LR.prob.threhsold = NULL,
  signature.col.aggregate = NULL,
  DEA.method = DEA.METHODS,
  DEA.data = NULL,
  sample.weights = NULL,
  folds.save.folder = NULL,
  file.CV = NULL,
  seed = 1234
)
}
\arguments{
\item{x}{data.frame or matrix; The input matrix corresponding to the feature 
space only. additional covariates are found in `y`. Rows = samples, 
Columns = features}

\item{y}{data.frame or matrix; The output matrix. 
Columns = c(y.label, y.covariates, y.sample)}

\item{CV.k}{numerical: The number of folds
Default = 5}

\item{CV.y.leave.1.out.column}{vector; If not NULL, vector to create Leave-1-out category CV}

\item{CV.replicates}{numerical; The number of replicates. 
At each replicate, the seed is increased by 1}

\item{CV.sampling.method}{character; The sampling method to balance the folds. 
if "under", Undersampling is performed to balance the categories in every fold
if "upper", Uppersampling is performed to balance the categories in every fold
if "none", no balancing sampling is performed. 
Default = "under"}

\item{y.label}{character; The column of `y` corresponding to the binary output}

\item{y.covariates}{character vector; The columns of `y` corresponding to the covariates}

\item{y.sample}{character; The column of `y` corresponding to the samples for when method = "signature"}

\item{method}{character; The prediction method. Possible values are: "LR": 
Logistic regression or "signature" signature-score approach.}

\item{hyperparams}{list: The hyper-parameter list. 
if method = "LR": Elements of list are 
  - "alpha" (L2-regularization): Vector of alpha values (0-1)
  - "lambda" (L1-regularization): Vector of lambda values (0-1)
if method = "signature": Elements of list are 
  - "signature.lengths": Vector of signature lengths
  - "signature.side": Vector of siganture sides ("up", "down" or "both")
  - "signature.rm.regex": Vector of rm.regex (regex to remove genes from signature)
  - "signature.methods": Vector of signature score methods}

\item{data.trans.method}{character: The data transforamtion method. 
Possible values are "pca" 
Default = "none" i.e. no transformation.}

\item{pca.explained.var.threshold}{numeric: If not NULL, apply this threshold 
to select PCs explaining more variance than the threhsold}

\item{rm.corr.features}{logical: Do we remove the mutually correlated feature 
while keeping the ones that best correlate with the outcome? 
default = F}

\item{DA.method}{character; The discriminant analysis (DA) method. 
Possible values are "wilcox"
Default = "none" i.e. no discriminant analysis}

\item{DA.p.adjust.method}{character; The p-value adjust method for the 
discriminant analysis. 
Default = "fdr"}

\item{DA.p.val.threshold}{numeric; The p-val significant threshold.}

\item{DA.event.per.variable, }{numerical; The number of event per variable threshold
A common EPV is 10 meaning that we need 10 event per variable. Here we use it as 
a way to select a fewer amount of variables for a fixed amount of samples. 
Default = NULL (no selection on EPV)}

\item{LR.prob.threhsold}{numeric; The probability threshold for the logitic regression}

\item{signature.col.aggregate}{character; Column in y to combine cells into pseudobulk. 
If NULL, no pseudobulk integration
Default = NULL}

\item{DEA.method}{character; The Differential-expression-analylsis method (see RunDEA())
default = "wilcox" from the Seurat::FindMarker() function}

\item{DEA.data}{data.frame or matrix; The input matrix for the DEA. 
Rows = samples, Columns = features. If NULL, the DEA.data is x
Default = NULL}

\item{sample.weights}{data.frame (rows = observation, column = weight); The 
observation weights data.frame. 
Default = NULL i.e. 1 for each observation}

\item{folds.save.folder}{character; Folder to save the fold model information. 
If null doesn't save anything. if not null, a folder called "NCV_", fold.out.name, "_", fold.in.name
will be created to store the model information in. 
default = NULL}

\item{file.CV}{character; If not null, save fold information to this file
Default = NULL}

\item{metric.to.optimize}{character; metric to optimize. 
Default = mcc}
}
\value{
data.frame summarizing the Nested-cross-validation
}
\description{
Home Cross validation function
}
