% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preparation.R
\name{DiscrAnalysisBinary}
\alias{DiscrAnalysisBinary}
\title{Discriminant Analysis Function}
\usage{
DiscrAnalysisBinary(
  x,
  y,
  y.label,
  y.sample = NULL,
  y.covariates = NULL,
  y.aggregate = NULL,
  method = DEA.METHODS,
  p.adjust.method = P.ADJUST.METHODS,
  p.val.threshold = 0.05,
  event.per.variable = NULL
)
}
\arguments{
\item{x}{data.frame; The input feature space data.}

\item{y}{categorical vector; The binary output.}

\item{method}{character; The discriminant analysis method. Possible value is "wilcox".}

\item{p.adjust.method}{character; The name of the method to adjust the p-value.
Possible values are methods in `stats::p.adjust()`, such as "bonferroni", "fdr" (deafault), 
"holm", "hochberg", "hommel", "BH", "BY", "none".}

\item{p.val.threshold}{numeric; The p-value significance threshold. Default = 0.05.}

\item{event.per.variable}{numeric; The number of events per variable threshold.
A common EPV is 10, meaning that we need 10 events per variable. Here, it is used 
to select fewer variables for a fixed number of samples. 
Default = NULL (no selection based on EPV).}
}
\value{
data.frame; The results from the discriminant analysis.
  - row names = features
  - column names = "p_val" (numeric), "p_val_adj" (numeric), and "Significant" (logical)
}
\description{
Computes discriminant analysis of a feature space for comparing two classes in a categorical variable.
The implemented discriminant analysis method is Wilcox rank sum test.
}
