% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preparation.R
\name{FeatureTrans}
\alias{FeatureTrans}
\title{Feature space transformation function}
\usage{
FeatureTrans(
  x.train,
  x.test = NULL,
  y.train.lab = NULL,
  method = FEATURE.TRANS.METHODS,
  explained.var.threshold = NULL
)
}
\arguments{
\item{x.train}{data.frame or matrix; The testing input data. Apply (predict) the trained PCA on it}

\item{method}{character; The data transformation method.
possible values are : pca
Default = pca}

\item{explained.var.threshold}{numerical; If method = "pca". Threshold to 
select the PCs with explained variable above the threshold
Default = NULL (no threshold)}
}
\value{
list of traingin and testing data.frames of transformed data
}
\description{
Function to transforme the feature space into another feature space. Example 
with PCA
}
