% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_evalution.R
\name{GetAccuracyMetrics}
\alias{GetAccuracyMetrics}
\title{Get binary accuracy metric}
\usage{
GetAccuracyMetrics(
  metrics = EVALUATION.METRICS,
  ground_truth = NULL,
  preds = NULL,
  tp = NULL,
  fp = NULL,
  tn = NULL,
  fn = NULL
)
}
\arguments{
\item{metrics}{character; The accuracy metrics. can be one or multiple metrics 
in "accuracy", "mcc", "F1", "kappa"}

\item{ground_truth}{character vector of ground truths}

\item{preds}{character vector of predictions}

\item{tp}{numerical; True Positives}

\item{fp}{numerical; False Positives}

\item{tn}{numerical; True Negatives}

\item{fn}{numerical; False Negatives}
}
\value{
list; list of accuracy measurements
}
\description{
Function to get a list of binary accuracy metrics
}
\details{
The user can provide ground_truth and preds and the function will compute tp, 
tn, fp, fn using the function GetMetricsBinary(). 
The user can also only provide tp, tn, fp, fn directly
}
