% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_evalution.R
\name{GetAUC}
\alias{GetAUC}
\title{Get AUC metrics}
\usage{
GetAUC(ground_truth, scores)
}
\arguments{
\item{ground_truth}{vector of ground truths.}

\item{scores}{numerical vector of scores}
}
\value{
numerical; The AUC
}
\description{
Function to get the AUC metric
}
