% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross_Validation.R
\name{GetBestHyperparams}
\alias{GetBestHyperparams}
\title{Get the Best Hyper-parameters for Nested Cross Validation}
\usage{
GetBestHyperparams(
  CV.res,
  method = c("LR", "signature"),
  CV.type = c("CV", "NCV"),
  outer.col = "outer",
  selection.metric = EVALUATION.METRICS
)
}
\arguments{
\item{CV.res}{data.frame; Nested-Cross-Validation results as returned by .CrossValidationLoop() or .NestedCrossValidationLoop().}

\item{method}{character; The method of the origin model. Possible values are
"LR" (default), "signature".}

\item{CV.type}{character; The name of the cross-validation type. Possible 
values are "CV" (default) or "NCV".}

\item{outer.col}{character; The name of the outer-fold column in CV.res.}

\item{selection.metric}{character; Selection metrics. Possible values are 
"mcc" (default), "accuracy", "F1", "kappa", "auc", "sensitivity", "specificity", "PPV", "NPV".}
}
\value{
data.frame; Data frame summarizing the best hyper-parameters for each outer fold.
}
\description{
Function to get the best hyper-parameters for the nested cross validation. 
It can compute all hyperparameters in a list. 
It can simultaneously get the best hyper-parameters for MCC, accuracy, F1, and kappa.
}
