% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{GetClonePred}
\alias{GetClonePred}
\title{Get clone-wise prediction}
\usage{
GetClonePred(
  pred.df,
  score.col = "score",
  pred.col = "pred",
  clone.col = "CloneId"
)
}
\arguments{
\item{pred.df}{data.frame; The prediction data-frame with 3 required columns: 
- the score column (score.col)
- the prediction column (pred.col)
- The clone-id column (clone.col)}

\item{score.col}{character; The numerical score column name in pred.df
Default = "score"}

\item{pred.col}{character; The boolean pred column name in pred.df
Default = "pred"}

\item{clone.col}{character; The character clone-id column in pred.df
Default = "CloneId"}
}
\value{
data.frame; Rows = clones id. Cols = "score.clone" and "pred.clone" and number of cells
}
\description{
Get clone wise prediction from cell-wise TRT predictions.
}
