% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_evalution.R
\name{GetMccBinary}
\alias{GetMccBinary}
\title{Get mcc metric}
\usage{
GetMccBinary(
  ground_truth = NULL,
  preds = NULL,
  tp = NULL,
  fp = NULL,
  tn = NULL,
  fn = NULL
)
}
\arguments{
\item{ground_truth}{character vector of ground truths}

\item{preds}{character vector of predictions}

\item{tp}{numerical; True Positives}

\item{fp}{numerical; False Positives}

\item{tn}{numerical; True Negatives}

\item{fn}{numerical; False Negatives}
}
\value{
numerical; The mcc
}
\description{
Function to get the mcc metric.
}
\details{
The user can provide ground_truth and preds and the function will compute tp, 
tn, fp, fn using the function GetMetricsBinary(). 
The user can also only provide tp, tn, fp, fn directly
}
