% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_evalution.R
\name{GetMetricsBinary}
\alias{GetMetricsBinary}
\title{Get binary predictive metric}
\usage{
GetMetricsBinary(ground_truth, preds)
}
\arguments{
\item{ground_truth}{character vector of ground truths}

\item{preds}{character vector of predictions}
}
\value{
list; list composed of the number of TP, TP, FP & FN
}
\description{
Function to get TP, TN, FP & FN
}
