% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{GetModelPrediction}
\alias{GetModelPrediction}
\title{Get Models perdiction}
\usage{
GetModelPrediction(
  data.input,
  data.output,
  y.label,
  path.folder,
  method,
  hyperparameters,
  data.trans.method,
  DEA.file.name = "dea.res.rds",
  signature.list = NULL,
  signature.x.threshold = 0,
  signature.score.scale = T,
  signature.score.scale.params = NULL,
  LR.model.file.name = NULL,
  LR.prob.threhsold = NULL
)
}
\arguments{
\item{data.input}{data.frame or matrix; The input matrix corresponding to the feature 
space only. Additional covariates are found in `y`. Rows = samples, Columns = features}

\item{data.output}{data.frame or matrix; The output matrix. 
Columns = c(y.label, y.covariates, y.sample)}

\item{y.label}{character; The column of `y` corresponding to the binary output}

\item{path.folder}{character; Folder to retieve the model / data / info
 @param method character; The prediction method. Possible values are: "LR": 
Logistic regression or "signature" signature-score approach.}

\item{data.trans.method}{character: The data transforamtion method. 
Possible values are "pca" 
Default = "none" i.e. no transformation.}

\item{DEA.file.name}{character; The name of the DEA file 
Default = "dea.res.rds"}

\item{signature.list}{list; The siganture list
Default = NULL}

\item{signature.x.threshold}{numerical; The signature score threhsold
Default = 0}

\item{LR.model.file.name}{character; The name of the LR file. 
If NULL, generate the file name generated by LRCrossValidation()
Default = NULL}

\item{LR.prob.threhsold}{numeric; The probability threshold for the logitic regression
Default = 0.5}

\item{hyperparams}{list: The hyper-parameter list. 
if method = "LR": Elements of list are 
  - "alpha" (L2-regularization): Vector of alpha values (0-1)
  - "lambda" (L1-regularization): Vector of lambda values (0-1)
if method = "signature": Elements of list are 
  - "signature.lengths": Vector of signature lengths
  - "signature.side": Vector of siganture sides ("up", "down" or "both")
  - "signature.rm.regex": Vector of rm.regex (regex to remove genes from signature)
  - "signature.methods": Vector of signature score methods
  - "signature.selection.method": How to select genes from DEA res using the log-Fold-Change ("logFC") or p-values ("pval")}
}
\value{
data.frame; Rows = barcode names in x. Cols = "score" and "pred"
}
\description{
Home function to train and test a model (LR or signature )
}
