% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_method.R
\name{GetPredictionSignature}
\alias{GetPredictionSignature}
\title{Get signature prediction from a given model}
\usage{
GetPredictionSignature(
  x,
  path.folder,
  DEA.file.name = "dea.res.rds",
  hyperparameters = DEFAULT.SIGNATURE.HYPERPARAMS,
  score.threshold = 0,
  score.to.scale = T,
  score.scale.params = NULL
)
}
\arguments{
\item{x}{Seurat; The Seurat input data}

\item{path.folder}{character; Folder to retrieve the model information.}

\item{DEA.file.name}{character; The DEA result file name}

\item{hyperparameters}{list; The signature hyperparmeters
Elements of list are 
  - "signature.lengths": Vector of signature lengths
  - "signature.side": Vector of siganture sides ("up", "down" or "both")
  - "signature.rm.regex": Vector of rm.regex (regex to remove genes from signature)
  - "signature.methods": Vector of signature score methods
  - "signature.selection.method": How to select genes from DEA res using the log-Fold-Change ("logFC") or p-values ("pval")}

\item{score.threshold}{numerical; The signature score threshold to binarize the score into positive and negative predictions}
}
\value{
data.frame; Rows = barcode names in x. Cols = "score" and "pred"
}
\description{
Function to get the signature model prediction when given the model.
}
