% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_functions.R
\name{getPseudoBulk}
\alias{getPseudoBulk}
\title{get Pseudo Bulk}
\usage{
getPseudoBulk(data, colData, col.aggregate)
}
\arguments{
\item{data}{matrix, assay with rows = genes and columns = observations (single-cells)}

\item{colData}{data.frame; columns = col data, rows = observations (single-cells)}

\item{col.aggregate}{character; aggregation column in colData}
}
\value{
list of two elements
- data = aggregated data (matrix)
- colData = aggreagted colData (data.frame)
}
\description{
Function to compute the pseudobulk information by combining information from single cells
}
