% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_method.R
\name{GetSignatureAccuracy}
\alias{GetSignatureAccuracy}
\title{Get Signature Accuracy}
\usage{
GetSignatureAccuracy(x, y, th)
}
\arguments{
\item{x}{numerical predictions}

\item{y}{logical ground truth}

\item{th}{numerical threhsold to apply on x to get logical predictions}
}
\value{
accuracy for x for a threshold th
}
\description{
Function that compute accuracy based on a threshold
}
