% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_method.R
\name{GetSignatureList}
\alias{GetSignatureList}
\title{Get signature lists from DE results and hyper-parameters}
\usage{
GetSignatureList(
  df,
  lengths,
  sides,
  remove_regex = NULL,
  gene.selection.methods = c("logFC"),
  pval_col = "padj",
  log2FC_col = "logFC",
  pval_limit = 0.05,
  log2FC_limits = c(0, 0)
)
}
\arguments{
\item{df}{data.frame:  DE results, row.names = gene names}

\item{lengths}{numerical vector; A list of signature lengths}

\item{sides}{character vector; A list of signature sides. Possible values are : 
up (only up-regulated genes), down (only down-regulated genes) and both (both 
up and down-regulated genes).}

\item{remove_regex}{character vector; A list of regex that defines genes names to remove.
if NULL or "none": No genes are removed. 
Default = NULL}

\item{gene.selection.methods}{character vector; signature gene selection method.
possible values are "logFC" (select from logFC) or "pval" (select from pval)
Default = c("logFC")}

\item{pval_col}{character: The name of the p adjusted column}

\item{log2FC_col}{character: The name of the Log Fold Change column}

\item{pval_limit}{numeric: The upper significance limit of the adjusted p-value}

\item{log2FC_limits}{vector of numeric: The lower and upper limits of the 
logFC column to define the up and down regulated genes.}
}
\value{
list; list of up and down regulated gene. names of the list are 
  - up = n up-regulated gene
  - down = n down regulated genes
}
\description{
Function to get signature.list from DE results. This functions returns a list 
of signatures got from GetSignature() with varying lengths or side or even 
with and without certain genes. The name of the resulting list corresponds to 
the hyperparameters mentionned above.
}
