% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature_score_function.R
\name{GetSignatureScore}
\alias{GetSignatureScore}
\title{Get signature score}
\usage{
GetSignatureScore(
  object,
  signature,
  ranks = NULL,
  assay = "RNA",
  slot = "data",
  method = SIGNATURE.SCORE.METHODS,
  scale = F
)
}
\arguments{
\item{object}{Seurat; Seurat single-cell object}

\item{signature}{list: signature as defined in GetSignature()}

\item{assay}{character; Assay in `object` to pull the slot data from. 
default = "RNA"}

\item{slot}{character; Slot to pull data from the `assay`. 
default = "data"}

\item{method}{character: The signature score method. 
possible methods are "average" (default), "UCell", "AUCell", "singscore"}

\item{scale}{logical: Weather to scale signature score or not.}
}
\value{
data.frame; The signature score data.frame. Row = cells and columns = names(signature)
}
\description{
Function to get signature score from different methods
}
