% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{GetTRTpred}
\alias{GetTRTpred}
\title{Get Models prediction}
\usage{
GetTRTpred(
  seurat.obj,
  seurat.assay = "RNA",
  seurat.slot = "scale.data",
  covariate.cols = NULL,
  y.label = NULL
)
}
\arguments{
\item{seurat.obj}{Seurat Object; RNA-sequencing single-cell object (required)}

\item{seurat.assay}{character; The name of the seurat assay. 
Default = "RNA"}

\item{seurat.slot}{character; The name of the seurat slot
Default = "scale.data"}

\item{covariate.cols}{list of character; covariates
Default = NULL}

\item{y.label}{character; The y-label name. 
Default = NULL}
}
\value{
data.frame; Cell-Wise TRT predictions 
Rows = barcode names in x. Cols = "score" and "pred"
}
\description{
Home function to apply TRTpred
}
