% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logistic_regression.R
\name{LRCrossValidation}
\alias{LRCrossValidation}
\title{Logistic Regression Cross validation function}
\usage{
LRCrossValidation(
  data.train,
  design.str,
  hyperparams,
  data.test = NULL,
  model.prob.threshold = NULL,
  sample.weights = NULL,
  folds.save.folder = NULL,
  save.model = F,
  save.pred = F,
  save.score = F
)
}
\arguments{
\item{data.train}{data.frame/matrix. The training input data with the dependent and independent variables.}

\item{design.str}{character; The design of the model}

\item{hyperparams}{list; The hyper-parameter list with two elements:
- "alpha" (L2-regularization): Vector of alpha values (0-1)
- "lambda" (L1-regularization): Vector of lambda values (0-1)}

\item{data.test}{data.frame/matrix. The testing input data with the dependent and independent variables
Default = NULL i.e. not computed}

\item{model.prob.threshold}{numerical; The model probability threshold
Default = 0.5}

\item{sample.weights}{data.frame (rows = observation, column = weight); The 
observation weights data.frame. 
Default = NULL i.e. 1 for each observation}

\item{folds.save.folder}{character; Folder to save the model information. 
If null doesn't save anything. If not null, it will save the following data into the folder: 
- res.fit.rds: glmnetUtils::glmnet() results
- y.train.pred.rds: y.train.pred data.frame (rows = cell name, col = prediction)
- y.test.pred.rds: y.test.pred data.frame (rows = cell name, col = prediction)
default = NULL}
}
\value{
data.frame; The dataframe summarizing the accuracy of the model for varying hyperparameters
}
\description{
The function to perform logistic regression within a cross-validation frame-work
}
