% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preparation.R
\name{PrepareData}
\alias{PrepareData}
\title{Prepare data for machine learning}
\usage{
PrepareData(
  x.train,
  y.train,
  y.label,
  x.test = NULL,
  y.test = NULL,
  DEA.data = NULL,
  method = c("LR", "signature"),
  data.trans.method = c("none", FEATURE.TRANS.METHODS),
  pca.explained.var.threshold = NULL,
  rm.corr.features = F,
  rm.corr.features.threhsold = 0.8,
  DA.method = c("none", DA.METHODS),
  DA.p.adjust.method = P.ADJUST.METHODS,
  DA.p.val.threshold = 0.05,
  DA.event.per.variable = NULL
)
}
\arguments{
\item{x.train}{data.frame or matrix; The input matrix corresponding to the feature 
space only. Additional covariates are found in `y`. Rows = samples, Columns = features}

\item{y.train}{data.frame or matrix; The output matrix. 
Columns = c(y.label, y.covariates, y.sample)}

\item{y.label}{character; The column of `y` corresponding to the binary output}

\item{x.test}{data.frame or matrix; The input data like x.train but for testing}

\item{y.test}{data.frame or matrix; The output matrix like y.train but for testing}

\item{DEA.data}{data.frame or matrix; The input matrix for the DEA. 
Rows = samples, Columns = features. If NULL, the DEA.data is x
Default = NULL}

\item{method}{character; The prediction method. Possible values are: "LR": 
Logistic regression or "signature" signature-score approach.}

\item{data.trans.method}{character: The data transforamtion method. 
Possible values are "pca" 
Default = "none" i.e. no transformation.}

\item{pca.explained.var.threshold}{numeric: If not NULL, apply this threshold 
to select PCs explaining more variance than the threhsold}

\item{rm.corr.features}{logical: Do we remove the mutually correlated feature 
while keeping the ones that best correlate with the outcome? 
default = F}

\item{rm.corr.features.threhsold}{numerical: The correlation threshold of rm.corr.features
Default = 0.8}

\item{DA.method}{character; The discriminant analysis (DA) method. 
Possible values are "wilcox"
Default = "none" i.e. no discriminant analysis}

\item{DA.p.adjust.method}{character; The p-value adjust method for the 
discriminant analysis. 
Default = "fdr"}

\item{DA.p.val.threshold}{numeric; The p-val significant threshold.}

\item{DA.event.per.variable, }{numerical; The number of event per variable threshold
A common EPV is 10 meaning that we need 10 event per variable. Here we use it as 
a way to select a fewer amount of variables for a fixed amount of samples. 
Default = NULL (no selection on EPV)}

\item{y.covariates}{character vector; The columns of `y` corresponding to the covariates}

\item{y.sample}{character; The column of `y` corresponding to the samples for when method = "signature"}
}
\value{
data.frame summarizing the Nested-cross-validation
}
\description{
Function which compiles the different steps to prepare the input data for some model. 
The steps are:
1. Data Transformation (data.trans.method)
2. Remove Correlated Features (rm.corr.features)
3. Discriminant analysis (DA.method)
}
\details{
Note that every step can be skipped by setting the corresponding parameters 
to "none", False, "none" respectively.
}
