% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ProcessConfigFile}
\alias{ProcessConfigFile}
\title{Process Config files}
\usage{
ProcessConfigFile(file, config)
}
\arguments{
\item{file}{Character; The config file path}

\item{config}{character; The config of the config file (see config::get())}
}
\value{
list with two elements: (1) config.file: List of the config parameters 
and (2) str.out a string summary of the parameters in the config filel
}
\description{
Function to process config file as in the config R library
}
