% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preparation.R
\name{RemoveMutuallyCorrFeatures}
\alias{RemoveMutuallyCorrFeatures}
\title{Remove mutually correlated features}
\usage{
RemoveMutuallyCorrFeatures(x, y, threshold = 0.8)
}
\arguments{
\item{x}{data.frame; The input feature space data}

\item{y}{categorical vector; The binary output}

\item{threshold}{numerical; Significant correlation threshold
default = 0.8}
}
\value{
data.frame. The results form the discriminant analysis. 
 - row.names = features
 - column names = "mutually_corr" (logical), "keep" (logical) & "abs_y_corr" (numerical)
}
\description{
This function outputs a data.frame indicating which features are mutually 
correlated and which one to remove.
}
