% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature_score_function.R
\name{run_scgsea}
\alias{run_scgsea}
\title{Run the scGSEA}
\usage{
run_scgsea(
  X,
  gene_sets,
  R = NULL,
  alpha = 0.25,
  scale = T,
  norm = F,
  single = T
)
}
\arguments{
\item{X}{matrix; The matrix. Columns = cells, rows = some features}

\item{gene_sets}{list of genes; the gene-sets}

\item{R}{rank-matrix as defined by matrixStats::colRanks
If provided, the function does not compute the rankings from X}
}
\value{
scGSEA results in data.frame format
}
\description{
Function to run the scGSEA
}
