% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_functions.R
\name{RunDEA}
\alias{RunDEA}
\title{Run Differential-Expression Analysis}
\usage{
RunDEA(
  object,
  col.DE_group,
  assay = NULL,
  slot = "data",
  method = c(SEURAT.METHODS, EDGER.METHODS, LIMMA.METHODS, DESEQ.METHODS),
  col.sample = NULL,
  col.covariate = NULL
)
}
\arguments{
\item{object}{Seurat; The Seurat input}

\item{col.DE_group}{character; The name of the column in meta.data 
corresponding to the differential expression analysis groupe}

\item{assay}{character; The Seurat assay. Default = NULL i.e. Seurat::DefaultAssay(object = object)}

\item{slot}{character; The slot of the assay. Default = "data"}

\item{method}{character; The Differential-Expression Analysis method. 
possible values are "wilcox" (default), "bimod", "roc", "t", "negbinom", 
"poisson", "LR", "DESeq2_Wald", "DESeq2_LRT", "edgeR_LRT", "edgeR_QFL", 
"limma_voom", "limma_trend"}

\item{col.sample}{character; The name of the column in meta.data 
corresponding to the sample information
Default = NULL}

\item{col.covariate}{character vector; The names of the column in meta.data 
corresponding to the covariates in the models
Default = NULL}
}
\description{
Function to run the differential expression analysis. This function calls the 
sub-function corresponding to the method asked.
}
