% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_functions.R
\name{RunedgeR}
\alias{RunedgeR}
\title{EdgeR DE function implementation}
\usage{
RunedgeR(
  mca,
  assay = NULL,
  slot = "data",
  method = c("QFL", "LRT"),
  col.DE_group = "DE_group",
  col.sample = "Patient",
  col.covariate = c("Patient")
)
}
\arguments{
\item{mca, }{Seurat: The Seurat object}

\item{assay, }{character. The Seurat assay. Default = Seurat::DefaultAssay(object = object)}

\item{slot, }{character. The Seurat assay's slot. Default = "data"}

\item{method, }{character: The limma method. Possible value are "QFL" (default) 
and "LRT"}

\item{col.DE_group, }{character: The colname in mca's meta.data defining the 
Differential Expression group. default = "DE_group". cannot be NULL}

\item{col.sample, }{character: The colname in mca's meta.data defining the 
sample information. default = "Patient". If NULL, every sample are in the same group}

\item{col.covariate, }{vector of character: A vector of colnames in mca's 
meta.data defining the covariates to add in the DE design. 
If NULL, no covariates are added. Default = c("Patient").}
}
\value{
data.frame. The DE result data.frame
}
\description{
Single-cell implementation of the EdgeR DE
}
