% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_functions.R
\name{Runlimma}
\alias{Runlimma}
\title{Limma DE function implementation}
\usage{
Runlimma(
  mca,
  assay = NULL,
  slot = "data",
  method = c("voom", "trend"),
  col.DE_group = "DE_group",
  col.sample = "Patient",
  col.covariate = c("Patient"),
  col.aggregate = NULL
)
}
\arguments{
\item{mca, }{Seurat: The Seurat object}

\item{assay, }{character. The Seurat assay. Default = Seurat::DefaultAssay(object = object)}

\item{slot, }{character. The Seurat assay's slot. Default = "data"}

\item{method, }{character: The limma method. Possible value are "voom" (default) 
and "trend"}

\item{col.DE_group, }{character: The colname in mca's meta.data defining the 
Differential Expression group. default = "DE_group". cannot be NULL}

\item{col.sample, }{character: The colname in mca's meta.data defining the 
sample information. default = "Patient". If NULL, every sample are in the same group}

\item{col.covariate, }{vector of character: A vector of colnames in mca's 
meta.data defining the covariates to add in the DE design. 
If NULL, no covariates are added. Default = c("Patient").}

\item{col.aggregate}{character; Column in mca@meta.data to combine cells into pseudobulk. 
If NULL, no pseudobulk integration
Default = NULL}
}
\value{
data.frame. The DE result data.frame
}
\description{
Single-cell implementation of the limma DE
}
