% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_functions.R
\name{RunSeurat}
\alias{RunSeurat}
\title{Seurat DE function implementation}
\usage{
RunSeurat(
  mca,
  assay = NULL,
  slot = "data",
  method = SEURAT.METHODS,
  col.DE_group = "DE_group"
)
}
\arguments{
\item{mca, }{Seurat: The Seurat object}

\item{assay, }{character. The Seurat assay. Default = Seurat::DefaultAssay(object = object)}

\item{slot, }{character. The Seurat assay's slot. Default = "data"}

\item{method, }{character: The limma method. Possible value are "wilcox" 
(default), "bimod", "roc", "t", "negbinom", "poisson" & "LR". The other
possible Seurat::FindMarkers methods are not implemented here.}

\item{col.DE_group, }{character: The colname in mca's meta.data defining the 
Differential Expression group. default = "DE_group". cannot be NULL}
}
\value{
data.frame. The DE result data.frame
}
\description{
Implementation of the Seurat DE
}
