% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature_score_function.R
\name{RunSigScoreAverage}
\alias{RunSigScoreAverage}
\title{Run the average signature score}
\usage{
RunSigScoreAverage(X, signature, scale = F)
}
\arguments{
\item{X}{matrix; The matrix. Columns = cells, rows = some features}

\item{signature}{list: signature as defined in GetSignature()}

\item{scale}{logical: Weather to scale signature score or not.}
}
\value{
data.frame; The signature score data.frame. Row = cells and columns = names(signature)
}
\description{
Function to run the average signature score
}
