% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature_score_function.R
\name{RunSigScoreUCell}
\alias{RunSigScoreUCell}
\title{Run the UCell signature score}
\usage{
RunSigScoreUCell(X, signature, ranks = NULL, scale = F)
}
\arguments{
\item{X}{matrix; The matrix. Columns = cells, rows = some features}

\item{signature}{list: signature as defined in GetSignature()}

\item{ranks}{rank-matrix as defined by UCell::StoreRankings_UCell(X). 
If provided, the function does not compute the rankings from X}

\item{scale}{logical: Weather to scale signature score or not.}
}
\value{
data.frame; The signature score data.frame. Row = cells and columns = names(signature)
}
\description{
Function to run the UCell signature score
}
