% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross_Validation.R
\name{CreateCVFolds}
\alias{CreateCVFolds}
\title{Create Cross-validation folds}
\usage{
CreateCVFolds(
  data.input,
  data.output,
  column.split,
  k.outer = 5,
  k.inner = 0,
  seed = 1234
)
}
\arguments{
\item{data.input}{data.frame; input data}

\item{data.output}{data.frame; output data}

\item{column.split}{vector of character or just one string; 
Columns in `data.output` that will be used to create "balanced" splitted data-sets}

\item{k.outer}{numerical: The number of outer folds
Default = 5}

\item{k.inner}{numerical: The number of inner folds. If 0, no inner fold is computed
Default = 0}

\item{seed}{numerical: The seed}
}
\value{
list of list of indexes creating the folds (see ? splitTools::create_folds())
}
\description{
Function to create cross validation folds using the `splitTools::create_folds()` function
}
