% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross_Validation.R
\name{CreateCVFolds}
\alias{CreateCVFolds}
\title{Create Cross-validation folds}
\usage{
CreateCVFolds(
  y,
  y.leave.1.out = NULL,
  k = 5,
  replicates = 1,
  sampling.method = c("under", "upper", "none"),
  seed = .Random.seed[1]
)
}
\arguments{
\item{y}{vector; the vector to create the CV folds from}

\item{y.leave.1.out}{vector; If not NULL, vector to create Leave-1-out category CV}

\item{k}{numerical: The number of outer folds
Default = 5}

\item{replicates}{numerical; The number of replicates. 
At each replicate, the seed is increased by 1. Default = 1}

\item{sampling.method}{character; The sampling method to balance the folds. 
if "under", Undersampling is performed to balance the categories in every fold
if "upper", Uppersampling is performed to balance the categories in every fold
if "none", no balancing sampling is performed. 
Default = "under"}

\item{seed}{numerical: The seed Default = .Random.seed[1]}
}
\value{
list containing the index of each sample belonging to the training and testing set for every fold
names of list = names of CV folds. values of list = list with two names 
- train: Training sample indexes
- test: Testing sample indexes
}
\description{
Function to create cross validation folds
}
