% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross_Validation.R
\name{CreateNestedCVFolds}
\alias{CreateNestedCVFolds}
\title{Create Nested-Cross-validation folds}
\usage{
CreateNestedCVFolds(
  y,
  y.leave.1.out = NULL,
  k.out = 5,
  k.in = 5,
  replicates = 1,
  sampling.method = c("under", "upper", "none"),
  seed = .Random.seed[1]
)
}
\arguments{
\item{y}{vector; the vector to create the CV folds from}

\item{y.leave.1.out}{vector; If not NULL, vector to create Leave-1-out category CV}

\item{k.out}{numerical: The number of outer folds
Default = 5}

\item{k.in}{numerical: The number of inner folds
Default = 5}

\item{replicates}{numerical; The number of replicates. 
At each replicate, the seed is increased by 1}

\item{sampling.method}{character; The sampling method to balance the folds. 
if "under", Undersampling is performed to balance the categories in every fold
if "upper", Uppersampling is performed to balance the categories in every fold
if "none", no balancing sampling is performed. 
Default = "under"}

\item{seed}{numerical: The seed Default = .Random.seed[1]}
}
\value{
list of list. first level = outer-folds, second level = inner folds. 
The content of the inner-fold list is the same as in CreateCVFolds()
}
\description{
Function to create Nested cross validation folds using the CreateCVFolds in this package.
See ? CreateCVFolds() for more details
}
