% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross_Validation.R
\name{CrossValidation}
\alias{CrossValidation}
\title{Cross Validataion}
\usage{
CrossValidation(
  x,
  y,
  folds.outer,
  y.label,
  y.covariates = NULL,
  y.sample = NULL,
  method = c("LR", "signature"),
  hyperparams = NULL,
  hyperparams.best.per.fold = NULL,
  data.trans.method = c("none", FEATURE.TRANS.METHODS),
  pca.explained.var.threshold = NULL,
  DA.method = c("none", DA.METHODS),
  DA.p.adjust.method = P.ADJUST.METHODS,
  DA.p.val.threshold = 0.05,
  DA.event.per.variable = NULL,
  LR.prob.threhsold = 0.5,
  DEA.method = DEA.METHODS,
  save.DEA.folder = NULL
)
}
\arguments{
\item{x}{data.frame or matrix; The input matrix corresponding to the feature 
space only. additional covariates are found in `y`. Rows = samples, 
Columns = features}

\item{y}{data.frame or matrix; The output matrix. 
Columns = c(y.label, y.covariates, y.sample)}

\item{folds.outer}{list; outer-folds definition list. See CreateCVFolds()}

\item{y.label}{character; The column of `y` corresponding to the binary output}

\item{y.covariates}{character vector; The columns of `y` corresponding to the covariates}

\item{y.sample}{character; The column of `y` corresponding to the samples for when method = "signature"}

\item{method}{character; The prediction method. Possible values are: "LR": 
Logistic regression or "signature" signature-score approach.}

\item{hyperparams}{list: The hyper-parameter list. 
If not NULL: Stand-alone cross-validation. If NULL, second part of NCV.
if method = "LR": Elements of list are 
  - "alpha" (L2-regularization): Vector of alpha values (0-1)
  - "lambda" (L1-regularization): Vector of lambda values (0-1)
if method = "signature": Elements of list are 
  - "signature.lengths": Vector of signature lengths
  - "signature.side": Vector of siganture sides ("up", "down" or "both")
  - "signature.rm.regex": Vector of rm.regex (regex to remove genes from signature)
  - "signature.methods": Vector of signature score method}

\item{hyperparams.best.per.fold}{data.frame; The best partameters for each 
fold found by Nested-cross-validation. This parameter is the output of 
GetBestOuterHyperparams()}

\item{data.trans.method}{character: The data transforamtion method. 
Possible values are "pca" 
Default = "none" i.e. no transformation.}

\item{pca.explained.var.threshold}{numeric: If not NULL, apply this threshold 
to select PCs explaining more variance than the threhsold}

\item{DA.method}{character; The discriminant analysis (DA) method. 
Possible values are "wilcox"
Default = "none" i.e. no discriminant analysis}

\item{DA.p.adjust.method}{character; The p-value adjust method for the 
discriminant analysis. 
Default = "fdr"}

\item{DA.p.val.threshold}{numeric; The p-val significant threshold.}

\item{DA.event.per.variable, }{numerical; The number of event per variable threshold
A common EPV is 10 meaning that we need 10 event per variable. Here we use it as 
a way to select a fewer amount of variables for a fixed amount of samples. 
Default = NULL (no selection on EPV)}

\item{LR.prob.threhsold}{numeric; The probability threshold for the logitic regression}

\item{DEA.method}{character; The Differential-expression-analylsis method (see RunDEA())
default = "wilcox" from the Seurat::FindMarker() function}

\item{save.DEA.folder}{character; Folder to save the DEA information in}
}
\value{
data.frame summarizing the Nested-cross-validation
}
\description{
Home Cross validation function.
This function can be used as a stand-alone Cross-validation function by providing the 
full `hyperparams` parameter. 
This function can also serve as the outer fold (second part) of a Nested-
cross-validation by providing the `hyperparams.best.per.fold` parameter.
}
