% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disciminant_Analysis_functions.R
\name{DiscrAnalysisBinary}
\alias{DiscrAnalysisBinary}
\title{Discriminant analysis function}
\usage{
DiscrAnalysisBinary(
  x,
  y,
  method = DA.METHODS,
  p.adjust.method = P.ADJUST.METHODS,
  p.val.threshold = 0.05
)
}
\arguments{
\item{x}{data.frame; The input feature space data}

\item{y}{categorical vector; The binary output}

\item{method}{character; The discriminant analysis method. Possible values are 
wilcox.}

\item{p.adjust.method, }{character; The name of the method to adjust the p-value.
possible values are methods in `stats::p.adjust()` i.e. "bonferroni", "fdr", 
"holm", "hochberg", "hommel", "BH", "BY", "none".
Default = "fdr"}

\item{p.val.threshold, }{numerical; The p-val significant threshold 
Default = 0.05}
}
\value{
data.frame. The results form the discriminant analysis. 
 - row.names = features
 - column names = "p_val" (numerical), "p_val_adj" (numerical) & "Significant" (logical)
}
\description{
This function compute the discriminant analysis of some feature space when 
comparing two classes in some categorical variable.
The discriminant analysis methods implemented are "wilcoxon rank sum test"
}
