% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preparation.R
\name{FeatureTrans}
\alias{FeatureTrans}
\title{Feature Space Transformation Function}
\usage{
FeatureTrans(
  x.train,
  x.test = NULL,
  y.train.lab = NULL,
  method = FEATURE.TRANS.METHODS,
  explained.var.threshold = NULL
)
}
\arguments{
\item{x.train}{data.frame or matrix; The training input data. Train the PCA on this data.}

\item{x.test}{data.frame or matrix; The testing input data. Apply (predict) the trained PCA on this data.}

\item{method}{character; The data transformation method. Possible values are: "pca".
Default = "pca".}

\item{explained.var.threshold}{numeric; If method = "pca", this threshold is used to 
select the PCs with explained variance above the threshold.
Default = NULL (no threshold).}
}
\value{
list; A list of training and testing data.frames of transformed data.
}
\description{
Function to transform the feature space into another feature space, e.g., with PCA.
}
