% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross_Validation.R
\name{GetBestHyperparams}
\alias{GetBestHyperparams}
\title{Get the best hyper-parameters for Nested cross validation}
\usage{
GetBestHyperparams(
  CV.res,
  method = c("LR", "signature"),
  CV.type = c("CV", "NCV"),
  outer.col = "outer",
  selection.metric = EVALUATION.METRICS
)
}
\arguments{
\item{CV.res}{data.frame; Nested-Cross-Validation results as returned by .CrossValidationLoop() or .NestedCrossValidationLoop()}

\item{method}{characters; The method of the origin model. possible values are
"LR" (default), "signature"}

\item{CV.type}{characters; The name of the cross-validation type. Possible 
values are "CV" (default) or "NCV"}

\item{outer.col}{character; The name of the outer-fold column in outer.col}

\item{selection.metric}{character; Selection metrics. Possible values are 
"mcc" (default), "accuracy", "F1", "kappa", "auc", "sensitivity", "specificity", "PPV", "NPV"}
}
\value{
data.frame; Data frame summarizing the best hyper-parameters for each outerfold
}
\description{
Function to get the best hyper-parameters for the nested cross validation. 
It can compute all hyperparameters in a list. 
It can simultaneously get the best hyper-parameters for mcc, accuracy, F1 and kappa.
}
