% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_method.R
\name{GetSignature}
\alias{GetSignature}
\title{Get signature from DE results}
\usage{
GetSignature(
  df,
  n_genes = 20,
  pval_col = "padj",
  log2FC_col = "logFC",
  pval_limit = 0.05,
  log2FC_limits = c(0, 0),
  remove_regex = NULL
)
}
\arguments{
\item{df}{data.frame:  DE results, row.names = gene names}

\item{pval_col}{character: The name of the p adjusted column}

\item{log2FC_col}{character: The name of the Log Fold Change column}

\item{pval_limit}{numeric: The upper significance limit of the adjusted p-value}

\item{log2FC_limits}{vector of numeric: The lower and upper limits of the 
logFC column to define the up and down regulated genes.}

\item{remove_regex}{character: Regular expression selecting genes to remove 
from the up and down gene-sets}
}
\value{
list; list of up and down regulated gene. names of the list are 
  - up = n up-regulated gene
  - down = n down regulated genes
}
\description{
Function to get signature from DE results
}
