% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logistic_regression.R
\name{LRCrossValidation}
\alias{LRCrossValidation}
\title{Logitic Regression Cross validation function}
\usage{
LRCrossValidation(
  data.train,
  design.str,
  hyperparams,
  data.test = NULL,
  model.prob.threshold = 0.5
)
}
\arguments{
\item{data.train}{data.frame/matrix. The training input data with the dependent and independent variables.}

\item{design.str}{character; The design of the model}

\item{hyperparams}{list; The hyper-parameter list with two elements:
- "alpha" (L2-regularization): Vector of alpha values (0-1)
- "lambda" (L1-regularization): Vector of lambda values (0-1)}

\item{data.test}{data.frame/matrix. The testing input data with the dependent and independent variables
Default = NULL i.e. not computed}

\item{model.prob.threshold}{numerical; The model probability threshold
Default = 0.5}
}
\value{
data.frame; The dataframe summarizing the accuracy of the model for varying hyperparameters
}
\description{
The function to perform logistic regression within a cross-validation frame-work
}
