% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_method.R
\name{SignatureCrossValidation}
\alias{SignatureCrossValidation}
\title{Signature Cross Validation}
\usage{
SignatureCrossValidation(
  data.train,
  y.label,
  data.test = NULL,
  y.sample = NULL,
  y.covariates = NULL,
  DEA.method = DEA.METHODS,
  DEA.data = NULL,
  signature.selection.method = c("logFC"),
  signature.lengths = c("20"),
  signature.sides = c("both"),
  signature.rm.regex = NULL,
  signature.methods = c("AUCell"),
  assay = "RNA",
  slot = "counts",
  col.aggregate = NULL,
  sample.weights = NULL,
  folds.save.folder = NULL,
  save.model = F,
  save.pred = F,
  save.score = F
)
}
\arguments{
\item{data.train}{Seurat; The Seurat training dataset}

\item{y.label}{character; The name of the meta.data column corresponding to the binary output}

\item{data.test}{Seurat; The Seurat testing dataset}

\item{y.sample}{character; The name of the meta.data column corresponding to the sample information (usefull for DESeq2, EdgeR and Limma)
Default = NULL i.e. no sample information}

\item{y.covariates}{vector of he name of the meta.data column corresponding to the models covariates
Default = NULL i.e. no covariates}

\item{DEA.method}{character; Differential-Expression-Analysis method name. 
possible values are possible values in runDEA()
Default = "wilcox" from Seurat::FindMarkers()}

\item{DEA.data}{data.frame or matrix; The input matrix for the DEA. 
Rows = samples, Columns = features. If NULL, the DEA.data is x
Default = NULL}

\item{signature.selection.method}{character vector; signatuire gene selection method.
possible values are "logFC" (select from logFC) or "pval" (select from pval)
Default = c("logFC")}

\item{signature.lengths}{numerical vector; Signature lengths}

\item{signature.sides}{character vector; Signature sides. Possible values ar
"up" (up-regulated genes), "down" (down-regulated genes) and "both" (both up 
and down-regulated genes)
Default = c("both")}

\item{signature.rm.regex}{character vector; A list of regex that defines 
genes names to remove. If genes are identified by the regex, they are removed.
if NULL or "none": No genes are removed. 
Default = NULL}

\item{signature.methods}{character vector; Signature score methods. 
Possible values are methods in GetSignatureScore()}

\item{assay}{character; The Seurat assay. Default = "RNA"}

\item{slot}{character; The Seurat slot. Default = "counts"}

\item{col.aggregate}{character; Column in data.train@meta.data to combine cells into pseudobulk. 
If NULL, no pseudobulk integration
Default = NULL}

\item{folds.save.folder}{character; Folder to save the model information. 
If null doesn't save anything. If not null, it will save the following data into the folder: 
- dea.res.rds: DEA results as in RunDEA()
- signature.score.train.rds: signature.score.train data.frame (rows = cell name, col = score for different hyperparmaters)
- signature.score.test.rds: signature.score.test data.frame (rows = cell name, col = score for different hyperparmaters)
- y.train.pred.rds: y.train.pred data.frame (rows = cell name, col = prediction)
- y.test.pred.rds: y.test.pred data.frame (rows = cell name, col = prediction)
default = NULL}
}
\value{
data.frame
}
\description{
Function to train and test a signature score model with a series of parameters
This function first compute the differential expression analysis results to create signatures. 
The signature score is then computed for the signatures and the function returns 
the accuracy predictions.
}
